+++ packages/react-router/lib/router/utils.ts
@@ -1386,10 +1386,7 @@ export const redirect: RedirectFunction = (url, init = 302) => {
   let headers = new Headers(responseInit.headers);
   headers.set("Location", url);
 
-  return new Response(null, {
-    ...responseInit,
-    headers,
-  });
+  return new Response(null, { ...responseInit, headers });
 };
 
 /**