+++ packages/react-router-node/CHANGELOG.md
@@ -1,5 +1,26 @@
 # `@react-router/node`
 
+## 7.1.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@7.1.1`
+
+## 7.1.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@7.1.0`
+
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@7.0.2`
+
 ## 7.0.1
 
 ### Patch Changes